+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Features:
 * Support KDE / LXQT wallpaper detection (Wallpaper, Linux)
 * Support QTerminal version & terminal font detection
+* Support MATE Terminal version & terminal font detection
 
 # 1.11.3